Package an unminified version of terser
This started in https://github.com/ahorek/terser-ruby/pull/4 but
couldn't see a reason why.

Minification makes it harder to debug issues and also makes it
much harder to review the gem changes on upgrade.

AFAIK this gem is only used with `exec-js` for which minification
doesn't impact performance much.